build: Enable -std=c99
So we can compile with C99 enabled; this requires checking that the
version of the compiler supports the -std=c99 command line switch, which
holds true for both GCC and CLang.

The warnings/errors we enable through the --enable-strict-flags
configure switch need to be amended so that we don't catch C99 valid
constructs like declaration after statement and VLAs.

[endlessm/eos-sdk#49]