Application model for wikipedia apps
Decompose the domain-specific apps into MVP design

We have a domain-wiki model that has a list of category models and
an image for that application.  We have a category model that has an
image, title, and a list of article models.  We have an article model
that has a human readable title and the url title for the database.

Application model now takes in json file, parses, it, and calls
addCategory() on each of the category objects

We moved the json reading code into the application model.
Now the model section of the app is essentially complete.
Category models are created by the app's initalization function,
which in turn create model objects for each article

[endlessm/eos-sdk#162]