Finalized models according to Philip's comments.
We have a domain-wiki model that has a list of category models and an image for that application.  We have a category model
that has an image, title, and a list of article models.  We have an article model that has a human readable title and the url
title for the database.

Application model now takes in json file, parses, it, and calls add category on each of the category objects
[endlessm/eos-sdk#162]

Dance dance dance dance
[endlessm/eos-sdk#162]