test commit [endlessm/eos-sdk#162]
Finished category and application model

We moved the json reading code into the application model. Now the model section of the app is essentially complete. Category models are created by the app's initalization function, which in turn create model objects for each article
[endlessm/eos-sdk#162]

Completed the view for GTK

I refactored more of the code so that it conforms with the MVP paradigm. All view related stuff is in the view files.
[endlessm/eos-sdk#171]

Added presenter

Added the presenter class to the domain specific app to complete MVP design
[endlessm/eos-sdk#175]