Created article list to look like designs
I modified the ArticleList widget to look like the designs. This meant modifying some of the css as well as creating a
new textButton widget, which displays a text label and an icon image, which will appear on a hover. Also added the
separators between the category info box and the article list

[endlessm/eos-sdk#177]