Add top bar edge finishing
Add a highlight and shadow, on the bottom of the top bar; two pixels
go over the top bar, and two pixels go over the app content.

Added a smoke test with a button that turns red on press, in order to
test that the two pixels on top of the app content do not catch input
events.

[endlessm/eos-sdk#197]