Restructured wikipedia webview to build HTML in memory
It was becoming too complicated trying to modify the index file on disk to account for relative paths.
I restructured the code so that the HTML is now built dynamically, in memory, and then saved to
a temp file, from which it is loaded into webkit. This way we can write all the paths based on the
current directory

[endlessm/eos-sdk#220]