Can add articles to the main category.
If there are articles for an app's main category, then it becomes clickable and acts like any other category.
If there are no articles for a category, it isn't cilckable and has no hover animations.
We no longer use the Title Label View since the main category is now a category button like any other.

[endlessm/eos-sdk#259]