Reorganized widget layout for the category page
We were encountering a bizarre bug where the textview would enlarge
when you click the button on the category page. It seems to be
due to the set_size_request call on the textview or containing
widget. To correct, I took out that set_size_request call and
instead subclassed the textview to override the get_preferred_width
function.
The button image is a placeholder till we get the right asset from
design.