Ensure a correct minimum height for EosFlexyGrid
The FlexyGrid cannot really have a fixed minimum height, since we don't
support width-for-height layout. This means that the FlexyGrid needs a
minimum height equal to its natural height - and we need to compute the
latter depending on the passed width.

This fixes the height handling when packing a EosFlexyGrid inside a
GtkScrolledWindow.

[endlessm/eos-sdk#1015]

[endlessm/eos-sdk#1015]