Remove screen-size test
This test does not apply anymore; previously we set the size of an
EosWindow explicitly to be the screen size, but now we just call
gtk_window_maximize(). A window manager is free to ignore this call, so
we cannot reliably test for it -- e.g. in Jenkins' xvfb environment.

[endlessm/eos-sdk#337]