Use AC_CONFIG_LINKS to copy file into builddir
Instead of a hacky blah: blah.in cp $< $@ rule.

Also, a stamp file turns out to be necessary for building documentation if
it is included in the distribution; if an up-to-date copy of the
documentation is unpacked into the source tree during make distcheck,
then it does not get remade in the build directory because it is
up-to-date. That causes make distcheck to fail later on, because it can't
copy the documentation from the build tree into the dist tree.

[endlessm/eos-sdk#340]