Fix wiki apps title category button sizing
Before this the size request of the title category button is set
based on the toplevels allocation. This broke when EosWindow's size
request was fixed.

Having some sort of container or bind constraint that allocated things
based on percentages is the eventual solution here. For now I used
a column homogeneous grid where each column was a percentage of the
allocation. A little gross, but works fine, and our wiki apps handle
resizing much better than they did before now
[endlessm/eos-sdk#376]

[endlessm/eos-sdk#376]