Force fill alignment of our AssetButton internals
Gtk 3.10 switched to baseline alignment for Gtk.Button label and
image widget. Which is not what we want. So I forall through the
internals and set the valign to fill (same as 3.8).

This isn't pretty, but unless Gtk provides some public API for
controlling our alignment, the only alternative is to use our
own box with our own internal widgets in AssetButton. Which would
be a lot of code duplicated from Gtk.Button

[endlessm/eos-sdk#427]