Set background color and import window border CSS
This sets the background color to a dark gray and imports the CSS
styling for the window border and shadow, because switching to
client-side decorations made the window background completely
transparent.

The resize grip is invisible because that's what the default in
eos-theme was, but the window can still be resized.

[endlessm/eos-sdk#28]