Move and realign category button
The category "go" arrow should never overlap the label text, so this
moves it out of its overlay and puts it in the same layer as the label.
The label now uses baseline alignment, and we can get rid of the hacky
extra padding in the CSS.

Currently, the innards of the category button look like this:
+---------------------------------------+
| Grid, hexpand: TRUE, halign: FILL     |
| +------------------+ +--------------+ |
| | Label            | | Arrow button | |
| | halign: START    | | halign: END  | |
| | valign: BASELINE | | valign: END  | |
| +------------------+ +--------------+ |
+---------------------------------------+
The arrow button does not have valign BASELINE, because that would
align it so it was even with the first line of the label if the label
was wrapped over two lines:

(GOOD)           (BAD)

HISTORY          HISTORY     >
& POLITICS  >    & POLITICS

Also, I have removed the "width-chars" property of the label, because,
as in the programming app in a commit which I cannot now find, when the
label wraps, the width-chars property makes it request a very odd size.

[endlessm/eos-sdk#505]