Show our internal widgets in eos-window
Users should not have be forced to call show all on the toplevel.
Calling show on all the widgets they create should be enough to
make things show on screen. Which means we need to call show on our
internals ourselves.

Added a test for this. Calls show on the window so will actually
make something pop on screen for a second. But so do other test we
have so I think thats OK
[endlessm/eos-sdk#535]