Always reserve space for arrow on article list
This keeps the content from jumping on article list mouseover. Used
the same opacity trick we use for the category button arrow. Did this
in js instead of css, and the image file was set in js and I wanted
to be consistent. But could also be done in css with state flags
[endlessm/eos-sdk#537]