Pagemanager subclasses GtkStack
Rather than keeping an internal stack. This saves us from a lot of
proxied container vfuncs, which had some bugs. Before we were
avoiding doing this for adding persistent overlays above the pages,
but don't think thats currently in the plans so we can take the
simpler approach of subclassing GtkStack
[endlessm/eos-sdk#985]