EosPageManager uses GtkStack properties directly
Use the visible-child and visible-child-name properties of GtkStack
in lieu of visible-page and visible-page-name. Use the
transition-duration, transition-type and name properties of GtkStack
instead of proxying in page manager with properties of that same name.
The page manger becomes nothing more than a stack with custom child
properties for use in EosWindow, and becomes a lot simpler.

This breaks the page manager api and will require changes in the
apps as well
[endlessm/eos-sdk#985]