Make topbar buttons no longer focusable
Set 'can-focus' properties to false for the buttons on the topbar
of the window (minimize, maximize, close). This was in response to
a bug in which the focus was being stolen by the maximize button
after clicking it and the eos-typing app was unable to regain,
now the behavior is consistent with all non-Endless applications.
[endlessm/eos-sdk#1043]