Add search box to SDK.
We reuse this same search box functionality in a
number of our apps so makes sense to stick it in the SDK.
Creates a search box that has autocompletion and provides
signals for when a search is triggered and when an
autocompletion entry is selected. Has its own default
styling which can be overridden.