Rename packages to include API version
Instead of eos-sdk-dev, we now have eos-sdk-0-dev. The packages all
"provide" their previous name for backwards compatibility. In addition,
eos-sdk is now called libendless-0-0 to comply with Debian best
practice.

This also adds an epoch to the debian/changelog, so that things don't
break because of backwards versioning.

In addition, this fixes a few Lintian warnings.

[endlessm/eos-sdk#418]
[endlessm/eos-sdk#418-debian]