Copy system theme for context menu styling
The current styling just left text as white, background as black
with no styling for hover and insensitive states, which looked
terrible. We've decided to just use the system theme instead

Unfortunately because there's no support for the :not selector
in css, there's no easy way to exempt our context menus from the
theming reset. So just duplicated the system theme for our context
menus for now
[endlessm/eos-sdk#617]