Change test in testWebActions
When calling app.quit(), the hold count on the application is ignored,
which causes undefined behavior when the application's window still
thinks it has a hold on the application and control is returned
to the MainLoop. This patch changes the way that applications are
quit in testWebActions by calling destroy() on the application's window,
which decrements the hold count on the application and implicitly causes
the application to close.

[endlessm/eos-sdk#1881]