TopbarNavButton works with no_show_all
Previously, if you created an Endless.TopbarNavButton with no_show_all
set to TRUE, and subsequently called show() on it, it would not become
visible, because the child widgets were still not shown. By showing them
at construct time, we solve this problem because they immediately become
visible when the TopbarNavButton widget itself is shown.

[endlessm/eos-sdk#2495]