Expose gettext() to client-side JS
This exposes the function set by webhelper.set_gettext() to the client-
side Javascript as a gettext() function, defined on the global window
object. This allows apps to translate messages that are generated
at runtime, not just messages in static HTML.

Some often-used JavaScriptCore operations can be turned into separate
functions, which we can put in a separate source file. This is in
anticipation of the next commit where we will define another function
property of the global object.

[endlessm/eos-sdk#291]