Add eos-html-extractor and m4 file
This is taken almost directly from the existing version in eos-english.
Cleanups to follow in subsequent commits. Previously the m4 code was in
two separate macros, but since they were much the same, I combined them
into one macro.

This also adds a very minimal test for eos-html-extractor; basically as
a very quick regression test for the cleanups to follow.

[endlessm/eos-sdk#3245]