eos-coverage: Add run_coverage.coverage "test"
This test automatically generates a coverage report at the
completion of all other tests. It exists as a pseudo-test which
depends on the completion of every other test. Effectively,
coverage reports are now generated automatically when
--enable-coverage is specified, though you can also run
make coverage-cobertura and make coverage-genhtml to update
them otherwise.

The burden on packages is that they must add .coverage to
TEST_EXTENSIONS and add run_coverage.coverage to TESTS.

[endlessm/eos-sdk#3303]