eos-coverage: Allow user to specify coverage data output directory
The default directory works for development machine cases, but for
distcheck, the data will be (or should be) cleaned away. CI systems
like jenkins need to preserve the data, so allow them to store it
somewhere else and only clean away the default directory on
the clean-coverage target.

Use --with-coverage-dir=DIRECTORY to specify it.

[endlessm/eos-sdk#3303]