Handle excess whitespace in strings
Whitespace between words and tags doesn't matter to HTML. Indeed, the
text in a translatable element may be formatted any way over any number
of lines, so we normalize all consecutive whitespace to be just one space
character and strip whitespace from the beginning and end of the strings.

This is so that translators are not confronted with strange newlines and
whitespace on Transifex.

[endlessm/eos-sdk#3291]