Handle HTML with embedded tags
When an element, such as <p>, has a name="translatable" attribute, we
also want to grab markup tags inside the element and translate them as
well.

For example, previously this HTML:

<p name="translatable">An embedded <b>tag</b> in a paragraph</p>

would result in the following string being extracted:

_("An embedded");

However, we want it to be:

_("An embedded <b>tag</b> in a paragraph");

This removes the use of BeautifulSoup from the eos-html-extractor script.
Unfortunately BeautifulSoup could have done this quite easily, but it
does not provide any line number information, which we need. Previously
in order to get the line numbers we also used html.parser from Python's
standard library, to augment the data we got from BeautifulSoup. However,
this issue required html.parser to do all the work that BeautifulSoup did
anyway, so there is no reason to use BeautifulSoup anymore.

[endlessm/eos-sdk#3291]