Process --with-coverage-dir flag even without coverage
Even if coverage is not requested, it is still turned on during "make
distcheck". In that case, any --with-coverage-dir flag passed to configure
should be respected.

In general, it's good practice not to let the _presence_ of a configure
flag depend on the value of another configure flag; only the actions taken
as a result of that flag.

[endlessm/eos-sdk#3540]