Remove custom spinner from CSS theme
We previously had a custom spinner because we reset the Adwaita spinner
but couldn't copy its code into our theme because it was GPL.
This changes our reset.css so that it doesn't apply to spinners.

Unfortunately, changing the reset selector from "EosWindow *" to
"EosWindow :not(.spinner)" increases its specificity, so it would
override things like ".menu" in endless-widgets.css. To avoid that, we
have to use ":not(GtkSpinner)" since element type selectors are less
specific than class selectors. This has the drawback that other widgets
with the .spinner class will still get reset, but we don't currently use
any.

[endlessm/eos-sdk#3634]