Don't treat locale getter as critical init section
We originally treated the get_locale method in the eoslicense API as a critical
initialization section. This no longer holds, since we could have different
locales for different types of licenses (e.g. creativecommons licenses are
translated, hence the locale will be "es", whereas GFDL licenses are not
translated, hence the locale will be "C").

[endlessm/eos-sdk#3471]