eostopbar: always allocate as if custom widgets were visible
This is an attempt to stop the top bar from jumping around on the
knowledge apps in composite mode.

This won't stop the top bar from jumping in size in all cases, if
you have different sized widgets forcing the top bar size to change
on different pages of the page manager, this won't really help you.
[endlessm/eos-sdk#3683]