Use constructed() for EosTopBar properties
Properly, object properties should not be set in init() in C, because
construct properties are later reset to their default values. Instead,
the constructed() vfunc should be used.

This doesn't matter for the current code, since none of the set
properties are construct properties, but I'm going to be doing something
in one of the following commits where it does matter.

[endlessm/eos-sdk#3782]