EosTopBar inherits from GtkHeaderBar
A lot of the functionality that we wrote in EosTopBar has been
implemented in GtkHeaderBar in the meantime. Using upstream code is
better because we don't have to maintain it, and can cut ~200 lines out
of our library.

Two discrepancies between the old and new EosTopBar is that the header
bar has now gained a right-click menu on which you can close the app (I
am leaving that in for now, as it's consistent with the other apps) and
there is a separator between the window minimize button and the credits
button. (The separator itself is invisible, but it adds another few
pixels of inter-widget space that weren't there before.)

[endlessm/eos-sdk#3782]