Shrink auto-complete result titles on small screen
When running on small screens, some auto-complete result titles do not fit in
the screen. This was most often observed in composite mode (720x480) and 800x600
resolutions.

Now, we ellipsize the cell renderer accordingly by checking the screen width,
based on the responsive system breakpoints defined by the Design team.

(Fernando and Martin).

[endlessm/eos-sdk#4016]