Connect to window-object-cleared earlier
If we connect to window-object-cleared only after we own the DBus name,
then we risk a race condition where a page is created and its window
object is already cleared by the time we connect to the signal, meaning
that the signal never fires for that page.

However, that on its own would introduce a different race condition,
where the JS code would call gettext() before the DBus connection is
obtained. Since JSCore doesn't allow you to create JS functions that
complete asynchronously, we do the simple but not-very-good thing: if
gettext() or ngettext() are called before there is a DBus connection, we
spin the main loop until the connection is acquired.

https://phabricator.endlessm.com/T11395