Record metric for unmaximized windows
When an EosWindow remains unmaximized for over 10 seconds, we record a
metric. The 10 seconds is to filter out presses of the unmaximize button
by accident, and because there is some weirdness in the window manager
around launching apps maximized.

This is for determining how important it is to implement various kinds of
responsive design in GTK.

https://phabricator.endlessm.com/T12233