SearchBox: Get correct index from selected match
When clicking on a result from the autocomplete popup, we should select
the first index of the tree path. (The list is a flat list, so there is
always only one index.)

This prevents a warning about a missing property of
[Symbol.toPrimitive], because we were indexing this._items with the
indices array (which is not a primitive) instead of its first item.

https://phabricator.endlessm.com/T21027