Add machine-readable diff output
If we're using eos-profile-diff in a CI infrastructure then we should
not be printing out data in a human readable format, as we may want to
interpret the output using scripts at a later point.

We can use JSON, instead, and allow redirecting the output to a file;
this way, we can store the output as an artifact, collect it, and parse
it later.