Remove deprecated gdk_screen_get_width/height
To achieve this, eos_is_composite_tv_screen is replaced with a
GdkMonitor equivalent: eos_is_composite_tv_monitor.

With this change, it is possible for EosWindow to better detect the
physical monitor it is displayed on and adapt accordingly. However,
the existing code in eoswindow.c does not make use of this
functionality.

https://phabricator.endlessm.com/T28284